+++ include/vsmc/utility/hdf5io.hpp
@@ -529,22 +529,22 @@ namespace internal
 
 template <typename TupleVectorType>
 inline void hdf5_tuple_vector_resize(
-    std::size_t n, TupleVectorType &vec, Position<0>)
+    std::size_t n, TupleVectorType &vec, Index<0>)
 {
     std::get<0>(vec).resize(n);
 }
 
 template <typename TupleVectorType, std::size_t Pos>
 inline void hdf5_tuple_vector_resize(
-    std::size_t n, TupleVectorType &vec, Position<Pos>)
+    std::size_t n, TupleVectorType &vec, Index<Pos>)
 {
     std::get<Pos>(vec).resize(n);
-    hdf5_tuple_vector_resize(n, vec, Position<Pos - 1>());
+    hdf5_tuple_vector_resize(n, vec, Index<Pos - 1>());
 }
 
 template <typename TupleVectorType, typename InputIter>
 inline void hdf5_tuple_vector_copy(
-    std::size_t n, TupleVectorType &vec, InputIter first, Position<0>)
+    std::size_t n, TupleVectorType &vec, InputIter first, Index<0>)
 {
     typedef typename std::iterator_traits<InputIter>::value_type tuple_type;
     typedef typename std::tuple_element<0, tuple_type>::type value_type;
@@ -556,7 +556,7 @@ inline void hdf5_tuple_vector_copy(
 
 template <typename TupleVectorType, typename InputIter, std::size_t Pos>
 inline void hdf5_tuple_vector_copy(
-    std::size_t n, TupleVectorType &vec, InputIter first, Position<Pos>)
+    std::size_t n, TupleVectorType &vec, InputIter first, Index<Pos>)
 {
     typedef typename std::iterator_traits<InputIter>::value_type tuple_type;
     typedef typename std::tuple_element<Pos, tuple_type>::type value_type;
@@ -564,29 +564,29 @@ inline void hdf5_tuple_vector_copy(
     InputIter ffirst = first;
     for (std::size_t i = 0; i != n; ++i, ++ffirst)
         dst[i] = std::get<Pos>(*ffirst);
-    hdf5_tuple_vector_copy(n, vec, first, Position<Pos - 1>());
+    hdf5_tuple_vector_copy(n, vec, first, Index<Pos - 1>());
 }
 
 template <typename TupleVectorType, typename TuplePtrType>
 inline void hdf5_tuple_vector_ptr(
-    const TupleVectorType &vec, TuplePtrType &ptr, Position<0>)
+    const TupleVectorType &vec, TuplePtrType &ptr, Index<0>)
 {
     std::get<0>(ptr) = std::get<0>(vec).data();
 }
 
 template <typename TupleVectorType, typename TuplePtrType, std::size_t Pos>
 inline void hdf5_tuple_vector_ptr(
-    const TupleVectorType &vec, TuplePtrType &ptr, Position<Pos>)
+    const TupleVectorType &vec, TuplePtrType &ptr, Index<Pos>)
 {
     std::get<Pos>(ptr) = std::get<Pos>(vec).data();
-    hdf5_tuple_vector_ptr(vec, ptr, Position<Pos - 1>());
+    hdf5_tuple_vector_ptr(vec, ptr, Index<Pos - 1>());
 }
 
 template <typename InputIter, typename... InputIters>
 inline void hdf5store_list_insert_tuple(std::size_t nrow,
     const std::string &file_name, const std::string &data_name,
     const std::tuple<InputIter, InputIters...> &first, const std::string *sptr,
-    Position<0>)
+    Index<0>)
 {
     typedef typename std::tuple_element<0,
         std::tuple<InputIter, InputIters...>>::type iter_type;
@@ -601,7 +601,7 @@ template <typename InputIter, typename... InputIters, std::size_t Pos>
 inline void hdf5store_list_insert_tuple(std::size_t nrow,
     const std::string &file_name, const std::string &data_name,
     const std::tuple<InputIter, InputIters...> &first, const std::string *sptr,
-    Position<Pos>)
+    Index<Pos>)
 {
     typedef typename std::tuple_element<Pos,
         std::tuple<InputIter, InputIters...>>::type iter_type;
@@ -611,7 +611,7 @@ inline void hdf5store_list_insert_tuple(std::size_t nrow,
     const dtype *data = data_ptr.get();
     hdf5store_list_insert<dtype>(nrow, file_name, data_name, data, *sptr);
     hdf5store_list_insert_tuple(
-        nrow, file_name, data_name, first, --sptr, Position<Pos - 1>());
+        nrow, file_name, data_name, first, --sptr, Index<Pos - 1>());
 }
 
 } // namespace vsmc::internal
@@ -647,7 +647,7 @@ inline void hdf5store_list(std::size_t nrow, const std::string &file_name,
     const std::string *sptr = vnames.get() + dim;
     hdf5store_list_empty(file_name, data_name, append);
     internal::hdf5store_list_insert_tuple(
-        nrow, file_name, data_name, first, --sptr, Position<dim - 1>());
+        nrow, file_name, data_name, first, --sptr, Index<dim - 1>());
 }
 
 /// \brief Store a Sampler in the HDF5 format
@@ -718,12 +718,12 @@ inline void hdf5store(const StateTuple<RowMajor, T, Types...> &state,
 
     std::tuple<std::vector<T>, std::vector<Types>...> data_vec;
     internal::hdf5_tuple_vector_resize(
-        state.size(), data_vec, Position<dim - 1>());
+        state.size(), data_vec, Index<dim - 1>());
     internal::hdf5_tuple_vector_copy(
-        state.size(), data_vec, state.data(), Position<dim - 1>());
+        state.size(), data_vec, state.data(), Index<dim - 1>());
 
     typename std::tuple<const T *, const Types *...> data_ptr;
-    internal::hdf5_tuple_vector_ptr(data_vec, data_ptr, Position<dim - 1>());
+    internal::hdf5_tuple_vector_ptr(data_vec, data_ptr, Index<dim - 1>());
 
     hdf5store_list(
         state.size(), file_name, data_name, data_ptr, vnames.data(), append);